Round Robin Matchups within the ELO range.
Bot1 is the bot that has gone the longest without a battle.
Bot2 is the bot that has gone the longest without a battle against Bot1

The $skipping_per is to create some randomness in the battle order to prevent predictability and to keep new bots and reactivated bots from having battles clumped together since they will by default by Bot2 until they've played all the bots.      Battles are not permanently skipped they are just delayed, keeping the overall structure and balance in tact for the most part.   100 value will put result in an endless loop, don't do it.